requirements.txt
python-dotenv~=1.0.0
Flask~=2.2.2
sentence-transformers
requests~=2.31.0
geopy~=2.4.1
gunicorn